Substract (All Dat BSF) - (All Zero DCM) at T=25C

+++ Calorimetry/Script/Extract-All-dat.sh
@@ -3,11 +3,22 @@
 for fname in *-exp.csv
 do
     name=`basename $fname -exp.csv`
-    echo "Extract [$name]"
-    ../Script/Extract-Plot-Data.py < $name-exp.csv > r.dat
-    gnuplot ../Script/Plot-Hf-svg.gnu
-    mv r.dat $name.csv
-    mv r.svg ../Plot/All/$name.svg
-done
+    if [[ ! -e "$name.csv" ]]
+    then
+	echo "Extract data from file: [$name]"
+	../Script/Extract-Plot-Data.py < $name-exp.csv > r.dat
+	gnuplot ../Script/Plot-Hf-svg.gnu
+	mv r.dat $name.csv
+	mv r.svg ../Plot/All/$name.svg
+    else
+	echo "Scip [$name]"
+    fi
 
-rm r.dat
+    if [[ ! -e "../../docs/All/$name.csv" ]]
+    then
+	echo "Copy data file: [$name]"
+	cp $name.csv ../../docs/All/$name.csv
+    else
+	echo "Save data file: [$name]"
+    fi
+done
